// src/utility/types/typeDPR.tsx
export type DPR = {
  dpr_import?: any
  map?: any
  project_name?: any
  location?: any
  reminder_emails?: any
  dpr_report_emails?: any
  access_key?: any
  year?: any
  execution_time?: any
  automation_email_trigger?: any
  profiles?: any
  ItemDesc?: any
  total_scope?: any
  dpr_item_desc_ids?: any
  project_end_date?: any
  project_start_date?: any
  actual_till_date?: any
  projects?: any
  plan_ftm?: any
  actual_ftm?: any
  today?: any
  dwg_avail?: any
  avatar?: any
  app_logo?: any
  module_id?:any
  all_sheet?:any
  module?:any
message?:any
work_pack?:any
sheetinfo?:any
read_status?:any
  app_name?: any
  sheet_id?:any
  address?: any
  token?: any
  data_date?: any
  original_import_file?: any
  profile_name?: any
  work_package?: any
  vendor?: any
  project?: any
  work_pack_id?: any
  vendor_id?: any
  project_id?: any
  permissions?: any
  se_name?: any
  roles?: any
  email?: any
  mobile_number?: any
  projectId?: any
  per_page_record?: any
  page?: any
  created_at?: any
  id?: any
  ids?: any
  user_id?: any
  sheet_name?: any
  name?: any
  cell_value?: any
  row_position?: any
  row_new_position?: any
  dpr_map?: [] | any
  manpower?: any
  type?: any
  file_path?: any
  reason?: any
  import?: any
  file?: any
  date?: any
  work_item?: any
  change_reason_for_plan_ftm?: any
  dpr?: any
  eventId?: any
  originalArgs?: any
  description?: any
  workPackId?: any
  unit_of_measure?: any
  man_power_type?: any
  formData?: any
  jsonData?: any
  role_id?: any
  password?: any
  confirm_password?: any
  status?: any
  error?: (e?: any) => void
  user_type_id?: any
  dpr_config_id?: any
  dpr_data?: any
  value?: any
  title?: any
  item_desc?: any
  orderno?: any
  dpr_config_ids?: any
  role?: any
  vendor_workpack?: any
  project_ids?: any
   uom?:any
}

export type DprReportList = {
  user_id?: any
  item_desc?: any
  vendor_id?: any
  project_id?: any
  work_pack_id?: any
  vender_id?: any
  dpr_config_id?: any
  dpr_manage_id?: any
  data_date?: any
  total_scope?: any
  actual_till_date?: any
  plan_ftm?: any
  actual_ftm?: any
  today?: any
  dwg_avail?: any
  manpower?: any
  work_item?: any
  change_reason_for_plan_ftm?: any
  created_at?: any
  updated_at?: any
  original_import_file?: any
  name?: any
  date?: any
  user?: any
  email?: any
  from_date?: any
  to_date?: any
}